request_required_params and errors for storages

+++ queue_api/storages.py
@@ -7,11 +7,13 @@ import json
 
 class StorageAddMessages(QueueEndpoint):
     request_required_params = [
-        'name',
-        'path',
+        'storage'
     ]
 
-    response_topic = 'ocular/server_name/storages/add/request'
+    response_topic = 'ocular/{server_name}/storages/add/request'
+
+    def __init__(self, server_name):
+        super().__init__(server_name=server_name)
 
     def handle_request(self, params):
         print('message received', flush=True)
@@ -19,8 +21,8 @@ class StorageAddMessages(QueueEndpoint):
         print('request uid', self.request_uid, flush=True)
         print('params', params, flush=True)
 
-        # if not self.check_request_params(params):
-        #     return
+        if not self.check_request_params(params):
+            return
 
         serializer_params = {
             'name': params['storage']['name'],
@@ -32,9 +34,8 @@ class StorageAddMessages(QueueEndpoint):
             storage = serializer.save()
             storage.save()
         else:
-            # raise Exception('serializer is wrong')
             errors = serializer.errors
-            msg = RequiredParamError('Validation error: "err"'.format(err=errors))
+            msg = RequestParamValidationError('Validation error: "err"'.format(err=errors))
             self.send_error_response(msg)
             return
 
@@ -47,7 +48,10 @@ class StorageDeleteMessage(QueueEndpoint):
         'id'
     ]
 
-    response_topic = 'ocular/server_name/storages/delete/request'
+    response_topic = 'ocular/{server_name}/storages/delete/request'
+
+    def __init__(self, server_name):
+        super().__init__(server_name=server_name)
 
     def handle_request(self, params):
         print('message received', flush=True)
@@ -55,16 +59,16 @@ class StorageDeleteMessage(QueueEndpoint):
         print('request uid', self.request_uid, flush=True)
         print('params', params, flush=True)
 
-        # if not self.check_request_params(params):
-        #     return
+        if not self.check_request_params(params):
+            return
 
         storage = Storage.objects.filter(id=params['id']).first()
         if storage:
             storage.delete()
         else:
             # raise Exception('storage does not exist')
-            msg = RequiredParamError('Validation error: "err"'.format(err=errors))
-            self.send_error_response(msg)
+            error = RequestParamValidationError('storage with id {id} not found'.format(id=params['id']))
+            self.send_error_response(error)
             return
 
         self.send_success_response()
@@ -73,7 +77,10 @@ class StorageDeleteMessage(QueueEndpoint):
 
 class StorageListMessage(QueueEndpoint):
 
-    response_topic = 'ocular/server_name/storages/list/request'
+    response_topic = 'ocular/{server_name}/storages/list/request'
+
+    def __init__(self, server_name):
+        super().__init__(server_name=server_name)
 
     def handle_request(self, params):
         print('message received', flush=True)
@@ -101,8 +108,13 @@ class StorageListMessage(QueueEndpoint):
 
 
 class StorageUpdateMessage(QueueEndpoint):
+    request_required_params = [
+        'storage'
+    ]
+    response_topic = 'ocular/{server_name}/storages/update/request'
 
-    response_topic = 'ocular/server_name/storages/update/request'
+    def __init__(self, server_name):
+        super().__init__(server_name=server_name)
 
     def handle_request(self, params):
         print('message received', flush=True)
@@ -110,12 +122,15 @@ class StorageUpdateMessage(QueueEndpoint):
         print('request uid', self.request_uid, flush=True)
         print('params', params, flush=True)
 
+        if not self.check_request_params(params):
+            return
+
         serializer_params = {
             'name': params['storage']['name'],
             'path': params['storage']['path']
         }
 
-        storage = Storage.objects.filter(id=params['id']).first()
+        storage = Storage.objects.filter(id=params['storage']['id']).first()
         if storage:
             serializer = StorageSerializer(data=serializer_params)
             if serializer.is_valid():
@@ -123,15 +138,15 @@ class StorageUpdateMessage(QueueEndpoint):
                 storage.path = params['path']
             else:
                 errors = serializer.errors
-                msg = RequiredParamError('Validation error: "err"'.format(err=errors))
+                msg = RequestParamValidationError('Validation error: "err"'.format(err=errors))
                 self.send_error_response(msg)
                 return
         else:
             # raise Exception('storage does not exist')
-            msg = RequiredParamError('Validation error: "err"'.format(err=errors))
-            self.send_error_response(msg)
+            error = RequestParamValidationError('storage with id {id} not found'.format(id=params['storage']['id']))
+            self.send_error_response(error)
             return
 
-        send_in_queue(self.response_topic, json.dumps(message))
+        self.send_success_response()
         return {'message sent'}
